+++ libs/environment/src/components/index.ts
@@ -2,5 +2,3 @@ export * from './network-loader';
 export * from './network-switcher';
 export * from './node-guard';
 export * from './node-switcher';
-export * from './node-switcher-2';
-export * from './node-switcher-dialog';
