--- libs/environment/src/components/node-switcher-2/index.tsx
@@ -1,1 +1,0 @@
-export * from './node-switcher';
