--- libs/environment/src/components/node-switcher-dialog/index.tsx
@@ -1,1 +1,0 @@
-export * from './node-switcher-dialog';
