+++ libs/environment/src/components/node-switcher/index.tsx
@@ -1,1 +1,2 @@
+export * from './node-switcher-dialog';
 export * from './node-switcher';
