+++ libs/environment/src/components/node-switcher/node-switcher-dialog.tsx
@@ -1,0 +1,16 @@
+import { Dialog } from '@vegaprotocol/ui-toolkit';
+import { NodeSwitcher } from './node-switcher';
+
+export const NodeSwitcherDialog = ({
+  open,
+  setOpen,
+}: {
+  open: boolean;
+  setOpen: (x: boolean) => void;
+}) => {
+  return (
+    <Dialog open={open} onChange={setOpen} size="medium">
+      <NodeSwitcher closeDialog={() => setOpen(false)} />
+    </Dialog>
+  );
+};
