+++ libs/environment/src/components/node-switcher/node-switcher.tsx
@@ -1,197 +1,247 @@
-import { useState } from 'react';
+import { useCallback, useState } from 'react';
 import { t } from '@vegaprotocol/react-helpers';
 import {
-  RadioGroup,
   Button,
+  ButtonLink,
   Input,
-  Link,
+  Loader,
   Radio,
+  RadioGroup,
 } from '@vegaprotocol/ui-toolkit';
-import { useEnvironment } from '../../hooks/use-environment';
-import { useNodes } from '../../hooks/use-nodes';
-import {
-  getIsNodeLoading,
-  getIsNodeDisabled,
-  getIsFormDisabled,
-  getErrorType,
-  getErrorByType,
-} from '../../utils/validate-node';
+import { useEnvironment } from '../../hooks';
 import { CUSTOM_NODE_KEY } from '../../types';
-import type { Configuration, NodeData, ErrorType } from '../../types';
+import { LayoutCell } from './layout-cell';
 import { LayoutRow } from './layout-row';
-import { NodeError } from './node-error';
-import { NodeStats } from './node-stats';
+import { ApolloWrapper } from './apollo-wrapper';
+import { RowData } from './row-data';
 
-type NodeSwitcherProps = {
-  error?: string;
-  config: Configuration;
-  initialErrorType?: ErrorType;
-  onConnect: (url: string) => void;
-};
-
-const getDefaultNode = (urls: string[], currentUrl?: string) => {
-  return currentUrl && urls.includes(currentUrl) ? currentUrl : undefined;
-};
-
-const getHighestBlock = (state: Record<string, NodeData>) => {
-  return Object.keys(state).reduce((acc, node) => {
-    const value = Number(state[node].block.value);
-    return value ? Math.max(acc, value) : acc;
-  }, 0);
-};
-
-export const NodeSwitcher2 = ({
-  config,
-  initialErrorType,
-  onConnect,
-}: NodeSwitcherProps) => {
-  const { VEGA_ENV, VEGA_URL } = useEnvironment();
-  const [networkError, setNetworkError] = useState(
-    getErrorByType(initialErrorType, VEGA_ENV, VEGA_URL)
+export const NodeSwitcher = ({ closeDialog }: { closeDialog: () => void }) => {
+  const { nodes, setUrl, status, VEGA_ENV, VEGA_URL } = useEnvironment(
+    (store) => ({
+      status: store.status,
+      nodes: store.nodes,
+      setUrl: store.setUrl,
+      VEGA_ENV: store.VEGA_ENV,
+      VEGA_URL: store.VEGA_URL,
+    })
   );
-  const [customNodeText, setCustomNodeText] = useState('');
-  const [nodeRadio, setNodeRadio] = useState(
-    getDefaultNode(config.hosts, VEGA_URL)
-  );
-  const { state, clients, updateNodeUrl, updateNodeBlock } = useNodes(config);
-  const highestBlock = getHighestBlock(state);
 
-  const customUrl = state[CUSTOM_NODE_KEY]?.url;
-
-  const onSubmit = (node: ReturnType<typeof getDefaultNode>) => {
-    if (node && state[node]) {
-      onConnect(state[node].url);
+  const [nodeRadio, setNodeRadio] = useState<string>(() => {
+    if (VEGA_URL) {
+      return VEGA_URL;
     }
-  };
+    return nodes.length > 0 ? '' : CUSTOM_NODE_KEY;
+  });
+  const [highestBlock, setHighestBlock] = useState<number | null>(null);
+  const [customUrlText, setCustomUrlText] = useState('');
 
-  const isSubmitDisabled = getIsFormDisabled(nodeRadio, VEGA_ENV, state);
+  const handleHighestBlock = useCallback((blockHeight: number) => {
+    setHighestBlock((curr) => {
+      if (curr === null) {
+        return blockHeight;
+      }
+      if (blockHeight > curr) {
+        return blockHeight;
+      }
+      return curr;
+    });
+  }, []);
 
-  const customNodeData =
-    nodeRadio &&
-    state[CUSTOM_NODE_KEY] &&
-    state[CUSTOM_NODE_KEY].url === customNodeText
-      ? state[nodeRadio]
-      : undefined;
-
-  const customNodeError = getErrorByType(
-    getErrorType(VEGA_ENV, customNodeData),
-    VEGA_ENV,
-    customUrl
-  );
+  let isDisabled = false;
+  if (nodeRadio === '') {
+    isDisabled = true;
+  } else if (nodeRadio === VEGA_URL) {
+    isDisabled = true;
+  } else if (nodeRadio === CUSTOM_NODE_KEY) {
+    if (!isValidUrl(customUrlText)) {
+      isDisabled = true;
+    }
+  }
 
   return (
-    <div className="text-black dark:text-white w-full lg:min-w-[800px]">
-      <NodeError {...(customNodeError || networkError)} />
-      <form
-        onSubmit={(event) => {
-          event.preventDefault();
-          onSubmit(nodeRadio);
-        }}
-      >
-        <p className="text-lg mt-4">
-          {t('Select a GraphQL node to connect to:')}
-        </p>
-        <div className="mb-2">
-          <div className="hidden lg:block">
-            <LayoutRow>
-              <div />
-              <span className="text-right">{t('Response time')}</span>
-              <span className="text-right">{t('Block')}</span>
-              <span className="text-right">{t('Subscription')}</span>
-            </LayoutRow>
-          </div>
+    <div>
+      <h3 className="uppercase text-xl calt text-center mb-2">
+        {t('Connected node')}
+      </h3>
+      {status === 'pending' ? (
+        <div className="py-8">
+          <p className="mb-4 text-center">{t('Loading configuration...')}</p>
+          <Loader size="large" />
+        </div>
+      ) : (
+        <div>
+          <p className="mb-2 text-sm text-center">
+            {t(
+              `This app will only work on ${VEGA_ENV}. Select a node to connect to.`
+            )}
+          </p>
           <RadioGroup
             value={nodeRadio}
-            onChange={(value) => {
-              setNodeRadio(value);
-              setNetworkError(null);
-            }}
+            onChange={(value) => setNodeRadio(value)}
           >
-            <div className="w-full">
-              {config.hosts.map((node, index) => (
-                <NodeStats
-                  key={index}
-                  data={state[node]}
-                  client={clients[node]}
+            <div className="hidden lg:block">
+              <LayoutRow>
+                <span>{t('Node')}</span>
+                <span className="text-right">{t('Response time')}</span>
+                <span className="text-right">{t('Block')}</span>
+                <span className="text-right">{t('Subscription')}</span>
+              </LayoutRow>
+              <div>
+                {nodes.map((node, index) => {
+                  return (
+                    <LayoutRow key={node} dataTestId="node-row">
+                      <ApolloWrapper url={node}>
+                        <RowData
+                          id={index.toString()}
+                          url={node}
+                          highestBlock={highestBlock}
+                          onBlockHeight={handleHighestBlock}
+                        />
+                      </ApolloWrapper>
+                    </LayoutRow>
+                  );
+                })}
+                <CustomRowWrapper
+                  inputText={customUrlText}
+                  setInputText={setCustomUrlText}
+                  nodes={nodes}
                   highestBlock={highestBlock}
-                  setBlock={(block) => updateNodeBlock(node, block)}
-                >
-                  <div className="break-all" data-testid="node">
-                    <Radio
-                      id={`node-url-${index}`}
-                      value={node}
-                      label={node}
-                      disabled={getIsNodeDisabled(VEGA_ENV, state[node])}
-                    />
-                  </div>
-                </NodeStats>
-              ))}
-              <NodeStats
-                data={state[CUSTOM_NODE_KEY]}
-                client={customUrl ? clients[customUrl] : undefined}
-                highestBlock={highestBlock}
-                setBlock={(block) => updateNodeBlock(CUSTOM_NODE_KEY, block)}
-              >
-                <div className="flex w-full mb-2">
-                  <Radio
-                    id={`node-url-custom`}
-                    value={CUSTOM_NODE_KEY}
-                    label={
-                      nodeRadio === CUSTOM_NODE_KEY || !!state[CUSTOM_NODE_KEY]
-                        ? ''
-                        : t('Other')
-                    }
-                  />
-                  {(customNodeText || nodeRadio === CUSTOM_NODE_KEY) && (
-                    <div
-                      data-testid="custom-node"
-                      className="flex items-center w-full gap-2"
-                    >
-                      <Input
-                        placeholder="https://"
-                        value={customNodeText}
-                        hasError={
-                          !!customNodeText &&
-                          !!(
-                            customNodeError?.headline ||
-                            customNodeError?.message
-                          )
-                        }
-                        onChange={(e) => setCustomNodeText(e.target.value)}
-                      />
-                      <Link
-                        aria-disabled={
-                          !customNodeText ||
-                          getIsNodeLoading(state[CUSTOM_NODE_KEY])
-                        }
-                        onClick={() => {
-                          setNetworkError(null);
-                          updateNodeUrl(CUSTOM_NODE_KEY, customNodeText);
-                        }}
-                      >
-                        {state[CUSTOM_NODE_KEY] &&
-                        getIsNodeLoading(state[CUSTOM_NODE_KEY])
-                          ? t('Checking')
-                          : t('Check')}
-                      </Link>
-                    </div>
-                  )}
-                </div>
-              </NodeStats>
+                  onBlockHeight={handleHighestBlock}
+                  nodeRadio={nodeRadio}
+                />
+              </div>
             </div>
           </RadioGroup>
+          <div className="mt-4">
+            <Button
+              fill={true}
+              disabled={isDisabled}
+              onClick={() => {
+                if (nodeRadio === CUSTOM_NODE_KEY) {
+                  setUrl(customUrlText);
+                } else {
+                  setUrl(nodeRadio);
+                }
+                closeDialog();
+              }}
+              data-testid="connect"
+            >
+              {t('Connect to this node')}
+            </Button>
+          </div>
         </div>
-        <div>
-          <Button
-            disabled={isSubmitDisabled}
-            fill={true}
-            type="submit"
-            data-testid="connect"
-          >
-            {t('Connect')}
-          </Button>
-        </div>
-      </form>
+      )}
     </div>
   );
+};
+
+interface CustomRowWrapperProps {
+  inputText: string;
+  setInputText: (text: string) => void;
+  nodes: string[];
+  highestBlock: number | null;
+  nodeRadio: string;
+  onBlockHeight: (blockHeight: number) => void;
+}
+
+const CustomRowWrapper = ({
+  inputText,
+  setInputText,
+  nodes,
+  highestBlock,
+  nodeRadio,
+  onBlockHeight,
+}: CustomRowWrapperProps) => {
+  const [displayCustom, setDisplayCustom] = useState(false);
+  const [error, setError] = useState<string | null>(null);
+  const showInput = nodeRadio === CUSTOM_NODE_KEY || nodes.length <= 0;
+
+  return (
+    <LayoutRow dataTestId="custom-row">
+      <div className="flex w-full mb-2">
+        {nodes.length > 0 && (
+          <Radio
+            id="node-url-custom"
+            value={CUSTOM_NODE_KEY}
+            label={nodeRadio === CUSTOM_NODE_KEY ? '' : t('Other')}
+          />
+        )}
+        {showInput && (
+          <div
+            data-testid="custom-node"
+            className="flex items-center w-full gap-2"
+          >
+            <Input
+              placeholder="https://"
+              value={inputText}
+              hasError={Boolean(error)}
+              onChange={(e) => {
+                setDisplayCustom(false);
+                setInputText(e.target.value);
+              }}
+            />
+            <ButtonLink
+              onClick={() => {
+                if (!isValidUrl(inputText)) {
+                  setError('Invalid url');
+                  return;
+                }
+                setError(null);
+                setDisplayCustom(true);
+              }}
+            >
+              {t('Check')}
+            </ButtonLink>
+          </div>
+        )}
+      </div>
+      {displayCustom ? (
+        <ApolloWrapper url={inputText}>
+          <RowData
+            id={CUSTOM_NODE_KEY}
+            url={inputText}
+            onBlockHeight={onBlockHeight}
+            highestBlock={highestBlock}
+          />
+        </ApolloWrapper>
+      ) : (
+        <>
+          <LayoutCell
+            label={t('Response time')}
+            isLoading={false}
+            hasError={false}
+            dataTestId="response-time-cell"
+          >
+            {'-'}
+          </LayoutCell>
+          <LayoutCell
+            label={t('Block')}
+            isLoading={false}
+            hasError={false}
+            dataTestId="block-height-cell"
+          >
+            {'-'}
+          </LayoutCell>
+          <LayoutCell
+            label={t('Subscription')}
+            isLoading={false}
+            hasError={false}
+            dataTestId="subscription -cell"
+          >
+            {'-'}
+          </LayoutCell>
+        </>
+      )}
+    </LayoutRow>
+  );
+};
+
+export const isValidUrl = (url?: string) => {
+  if (!url) return false;
+  try {
+    new URL(url);
+    return true;
+  } catch {
+    return false;
+  }
 };
