+++ libs/environment/src/components/node-switcher/node-swticher.spec.tsx
@@ -1,5 +1,5 @@
 import { fireEvent, render, screen, within } from '@testing-library/react';
-import { NodeSwitcherContainer } from './node-switcher';
+import { NodeSwitcher } from './node-switcher';
 import type { EnvStore } from '../../hooks';
 import { useEnvironment } from '../../hooks';
 import { Networks } from '../../types';
@@ -25,7 +25,7 @@ describe('NodeSwitcherContainer', () => {
       VEGA_ENV: Networks.TESTNET,
       nodes: [],
     });
-    render(<NodeSwitcherContainer closeDialog={jest.fn()} />);
+    render(<NodeSwitcher closeDialog={jest.fn()} />);
     expect(
       screen.getByText(Networks.TESTNET.toLowerCase())
     ).toBeInTheDocument();
@@ -45,7 +45,7 @@ describe('NodeSwitcherContainer', () => {
       VEGA_ENV: Networks.TESTNET,
       nodes,
     });
-    render(<NodeSwitcherContainer closeDialog={jest.fn()} />);
+    render(<NodeSwitcher closeDialog={jest.fn()} />);
     nodes.forEach((node) => {
       expect(
         screen.getByRole('radio', { checked: false, name: node })
@@ -85,7 +85,7 @@ describe('NodeSwitcherContainer', () => {
       VEGA_URL: selectedNode,
       nodes,
     });
-    render(<NodeSwitcherContainer closeDialog={jest.fn()} />);
+    render(<NodeSwitcher closeDialog={jest.fn()} />);
 
     nodes.forEach((node) => {
       expect(
@@ -112,7 +112,7 @@ describe('NodeSwitcherContainer', () => {
       nodes,
       setUrl: mockSetUrl,
     });
-    render(<NodeSwitcherContainer closeDialog={jest.fn()} />);
+    render(<NodeSwitcher closeDialog={jest.fn()} />);
     expect(
       screen.getByRole('button', { name: 'Connect to this node' })
     ).toBeDisabled();
@@ -152,7 +152,7 @@ describe('NodeSwitcherContainer', () => {
       nodes,
       setUrl: mockSetUrl,
     });
-    render(<NodeSwitcherContainer closeDialog={jest.fn()} />);
+    render(<NodeSwitcher closeDialog={jest.fn()} />);
     expect(
       screen.getByRole('button', { name: 'Connect to this node' })
     ).toBeDisabled();
